// src/api/types.gen.ts
export type AdjustedAmountDto = {
    adjustments?: Array<AmountAdjustmentDto> | null;
    withoutAdjustments?: AmountDto | null;
    adjustment?: AmountDto | null;
    value?: AmountDto | null;
};

export type AdjustedPriceDto = {
    adjustments?: Array<PriceAdjustmentDto> | null;
    withoutAdjustments?: PriceDto | null;
    adjustment?: PriceDto | null;
    value?: PriceDto | null;
};

export type AdjustedPriceWithPreviousAdjustmentsDto = {
    adjustments?: Array<PriceAdjustmentDto> | null;
    withoutAdjustments?: PriceDto | null;
    adjustment?: PriceDto | null;
    value?: PriceDto | null;
    previousAdjustments?: PriceDto | null;
    withPreviousAdjustments?: PriceDto | null;
    totalAdjustment?: PriceDto | null;
};
// ...
export type CountryDto = {
    id: string;
    code: string;
    name?: string | null;
    store?: StoreBasicDto | StoreDto | StoreNamedReferenceDto | StoreReferenceDto | null;
    taxCalculationMethod?: TaxCalculationMethodDto | TaxCalculationMethodNamedReferenceDto | TaxCalculationMethodReferenceDto | null;
    defaultCurrency?: CurrencyDto | CurrencyNamedReferenceDto | CurrencyReferenceDto | null;
    defaultPaymentMethod?: PaymentMethodDto | PaymentMethodNamedReferenceDto | PaymentMethodReferenceDto | PaymentMethodWithFeeDto | null;
    defaultShippingMethod?: ShippingMethodDto | ShippingMethodNamedReferenceDto | ShippingMethodReferenceDto | ShippingMethodWithRatesDto | null;
    regions?: Array<RegionDto | RegionNamedReferenceDto | RegionReferenceDto> | null;
    sortOrder?: number | null;
};
// ...
export type CurrencyDto = {
    id: string;
    code: string;
    name?: string | null;
    store?: StoreBasicDto | StoreDto | StoreNamedReferenceDto | StoreReferenceDto | null;
    culture?: string | null;
    symbol?: string | null;
    decimalPlaces?: number | null;
    decimalSeparator?: string | null;
    groupSeparator?: string | null;
    formatTemplate?: string | null;
    allowedCountries?: Array<AllowedCountryDto> | null;
    sortOrder?: number | null;
};

export type CurrencyNamedReferenceDto = {
    id: string;
    code: string;
    name?: string | null;
};

export type CurrencyReferenceDto = {
    id: string;
    code: string;
};
// ...
export type DiscountDto = {
    id: string;
    alias: string;
    name?: string | null;
    store?: StoreBasicDto | StoreDto | StoreNamedReferenceDto | StoreReferenceDto | null;
    type: DiscountType;
    codes?: Array<DiscountCodeDto> | null;
    rules?: DiscountRuleConfigDto | null;
    rewards?: Array<DiscountRewardConfigDto> | null;
    blockFurtherDiscounts?: boolean | null;
    blockIfPreviousDiscounts?: boolean | null;
    startDate?: string | null;
    expiryDate?: string | null;
    isActive?: boolean | null;
    status: DiscountStatus;
    sortOrder?: number | null;
};

export type DiscountNamedReferenceDto = {
    id: string;
    alias: string;
    name?: string | null;
};

export type DiscountReferenceDto = {
    id: string;
    alias: string;
};
// ...
export type GiftCardDto = {
    id: string;
    code: string;
    store?: StoreBasicDto | StoreDto | StoreNamedReferenceDto | StoreReferenceDto | null;
    currency?: CurrencyDto | CurrencyNamedReferenceDto | CurrencyReferenceDto | null;
    order?: OrderDto | OrderReferenceDto | null;
    originalAmount?: AmountDto | null;
    remainingAmount?: AmountDto | null;
    expiryDate?: string | null;
    createDate?: string | null;
    isActive?: boolean | null;
    status: GiftCardStatus;
    properties?: {
        [key: string]: string;
    } | null;
};

export type GiftCardReferenceDto = {
    id: string;
    code: string;
};
// ...
export type LocationDto = {
    id: string;
    alias: string;
    name?: string | null;
    store?: StoreBasicDto | StoreDto | StoreNamedReferenceDto | StoreReferenceDto | null;
    addressLine1?: string | null;
    addressLine2?: string | null;
    city?: string | null;
    region?: string | null;
    countryIsoCode?: string | null;
    zipCode?: string | null;
    type: LocationType;
    sortOrder?: number | null;
};

export type LocationNamedReferenceDto = {
    id: string;
    alias: string;
    name?: string | null;
};

export type LocationReferenceDto = {
    id: string;
    alias: string;
};

export enum LocationType {
    COMMERCIAL = 'Commercial',
    RESIDENTIAL = 'Residential'
}

export enum MeasurementSystem {
    METRIC = 'Metric',
    IMPERIAL = 'Imperial'
}

export type MediaContentDto = {
    id: number;
    key: string;
};

export type NotificationHeaderModel = {
    message: string;
    category: string;
    type: EventMessageTypeModel;
};

export type OrderAppliedDiscountCodeDto = {
    discount?: DiscountDto | DiscountNamedReferenceDto | DiscountReferenceDto | null;
    code?: string | null;
    isFulfilled?: boolean | null;
};

export type OrderBillingInfoDto = {
    addressLine1?: string | null;
    addressLine2?: string | null;
    city?: string | null;
    zipCode?: string | null;
    country?: CountryDto | CountryNamedReferenceDto | CountryReferenceDto | null;
    region?: RegionDto | RegionNamedReferenceDto | RegionReferenceDto | null;
    companyName?: string | null;
    companyTaxCode?: string | null;
    contact?: ContactDto | null;
};

export type OrderCustomerInfoDto = {
    firstName?: string | null;
    lastName?: string | null;
    email?: string | null;
    telephone?: string | null;
    customerReference?: string | null;
};

export type OrderDto = {
    id: string;
    orderNumber?: string | null;
    store?: StoreBasicDto | StoreDto | StoreNamedReferenceDto | StoreReferenceDto | null;
    cartNumber?: string | null;
    languageIsoCode?: string | null;
    currency?: CurrencyDto | CurrencyNamedReferenceDto | CurrencyReferenceDto | null;
    taxClass?: TaxClassDto | TaxClassNamedReferenceDto | TaxClassReferenceDto | null;
    taxRate?: number | null;
    orderStatus?: OrderStatusDto | OrderStatusNamedReferenceDto | OrderStatusReferenceDto | null;
    customer?: OrderCustomerInfoDto | null;
    billing?: OrderBillingInfoDto | null;
    payment?: OrderPaymentInfoDto | null;
    shipping?: OrderShippingInfoDto | null;
    transaction?: OrderTransactionInfoDto | null;
    discountCodes?: Array<OrderAppliedDiscountCodeDto> | null;
    discounts?: Array<DiscountDto | DiscountNamedReferenceDto | DiscountReferenceDto> | null;
    giftCards?: Array<GiftCardDto | GiftCardReferenceDto> | null;
    totalQuantity?: number | null;
    subtotalPrice?: AdjustedPriceWithPreviousAdjustmentsDto | null;
    totalPrice?: AdjustedPriceWithPreviousAdjustmentsDto | null;
    transactionAmount?: AdjustedAmountDto | null;
    orderLines?: Array<OrderLineDto> | null;
    properties?: {
        [key: string]: string;
    } | null;
    notes?: OrderNotesDto | null;
    tags?: Array<string> | null;
    createDate?: string | null;
    updateDate?: string | null;
    finalizedDate?: string | null;
    isFinalized?: boolean | null;
};

export type OrderLineDto = {
    id: string;
    productReference?: string | null;
    productVariantReference?: string | null;
    sku?: string | null;
    name?: string | null;
    quantity?: number | null;
    taxClass?: TaxClassDto | TaxClassNamedReferenceDto | TaxClassReferenceDto | null;
    bundleId?: string | null;
    orderLines?: Array<OrderLineDto> | null;
    properties?: {
        [key: string]: string;
    } | null;
    attributes?: Array<AttributeCombinationDto> | null;
    basePrice?: AdjustedPriceDto | null;
    unitPrice?: AdjustedPriceDto | null;
    taxRate?: number | null;
    totalPrice?: AdjustedPriceWithPreviousAdjustmentsDto | null;
};

export type OrderNotesDto = {
    customerNotes?: string | null;
    internalNotes?: string | null;
};

export type OrderPaymentInfoDto = {
    paymentMethod?: PaymentMethodDto | PaymentMethodNamedReferenceDto | PaymentMethodReferenceDto | PaymentMethodWithFeeDto | null;
    taxRate?: number | null;
    totalPrice?: AdjustedPriceDto | null;
};

export type OrderReferenceDto = {
    id: string;
    orderNumber?: string | null;
};

export type OrderRefundRequest = {
    refundAmount: number;
    orderlines?: Array<OrderlineRefundRequest> | null;
    restockProducts: boolean;
};

export enum OrderRoundingMethod {
    UNIT = 'Unit',
    LINE = 'Line',
    TOTAL = 'Total'
}

export type OrderShippingInfoDto = {
    addressLine1?: string | null;
    addressLine2?: string | null;
    city?: string | null;
    zipCode?: string | null;
    country?: CountryDto | CountryNamedReferenceDto | CountryReferenceDto | null;
    region?: RegionDto | RegionNamedReferenceDto | RegionReferenceDto | null;
    companyName?: string | null;
    companyTaxCode?: string | null;
    sameAsBilling?: boolean | null;
    contact?: ContactDto | null;
    shippingMethod?: ShippingMethodDto | ShippingMethodNamedReferenceDto | ShippingMethodReferenceDto | ShippingMethodWithRatesDto | null;
    shippingOption?: ShippingOptionDto | null;
    taxRate?: number | null;
    totalPrice?: AdjustedPriceDto | null;
};

export type OrderStatusDto = {
    id: string;
    alias: string;
    name?: string | null;
    store?: StoreBasicDto | StoreDto | StoreNamedReferenceDto | StoreReferenceDto | null;
    color?: string | null;
    sortOrder?: number | null;
};

export type OrderStatusNamedReferenceDto = {
    id: string;
    alias: string;
    name?: string | null;
};

export type OrderStatusReferenceDto = {
    id: string;
    alias: string;
};

export type OrderTransactionInfoDto = {
    transactionId?: string | null;
    authorizedAmount?: AmountDto | null;
    feeAmount?: AmountDto | null;
    baseCurrencyExchangeRate?: number | null;
    paymentStatus: PaymentStatus;
    refundedAmount?: AmountDto | null;
    refundableAmount?: AmountDto | null;
    transactionActivities: Array<TransactionActivityDto>;
};
// ...
export type StoreBasicDto = {
    id: string;
    alias: string;
    name?: string | null;
    measurementSystem: MeasurementSystem;
    sortOrder?: number | null;
};

export type StoreClaimsDto = {
    id: string;
    alias: string;
    name?: string | null;
    claims: Array<string>;
};

export type StoreDto = {
    id: string;
    alias: string;
    name?: string | null;
    measurementSystem: MeasurementSystem;
    sortOrder?: number | null;
    baseCurrency?: CurrencyDto | CurrencyNamedReferenceDto | CurrencyReferenceDto | null;
    defaultLocation?: LocationDto | LocationNamedReferenceDto | LocationReferenceDto | null;
    defaultCountry?: CountryDto | CountryNamedReferenceDto | CountryReferenceDto | null;
    defaultTaxClass?: TaxClassDto | TaxClassNamedReferenceDto | TaxClassReferenceDto | null;
    defaultOrderStatus?: OrderStatusDto | OrderStatusNamedReferenceDto | OrderStatusReferenceDto | null;
    errorOrderStatus?: OrderStatusDto | OrderStatusNamedReferenceDto | OrderStatusReferenceDto | null;
    pricesIncludeTax?: boolean | null;
    canRefundTransactionFee: boolean;
    useCookies?: boolean | null;
    cookieTimeout?: number | null;
    confirmationEmailTemplate?: EmailTemplateDto | EmailTemplateNamedReferenceDto | EmailTemplateReferenceDto | null;
    errorEmailTemplate?: EmailTemplateDto | EmailTemplateNamedReferenceDto | EmailTemplateReferenceDto | null;
    abandonedCartEmailTemplate?: EmailTemplateDto | EmailTemplateNamedReferenceDto | EmailTemplateReferenceDto | null;
    abandonedCartInactivityPeriod: number;
    abandonedCartLandingPageUrl?: string | null;
    cartNumberTemplate?: string | null;
    orderNumberTemplate?: string | null;
    orderRoundingMethod: OrderRoundingMethod;
    productPropertyAliases?: string | null;
    productUniquenessPropertyAliases?: string | null;
    giftCardCodeLength?: number | null;
    giftCardDaysValid?: number | null;
    giftCardCodeTemplate?: string | null;
    giftCardPropertyAliases?: string | null;
    giftCardActivationMethod: GiftCardActivationMethod;
    giftCardActivationOrderStatus?: OrderStatusDto | OrderStatusNamedReferenceDto | OrderStatusReferenceDto | null;
    defaultGiftCardEmailTemplate?: EmailTemplateDto | EmailTemplateNamedReferenceDto | EmailTemplateReferenceDto | null;
    allowedUsers?: Array<string> | null;
    allowedUserRoles?: Array<string> | null;
    logoImageUrl?: string | null;
    themeColor?: string | null;
};

export type StoreNamedReferenceDto = {
    id: string;
    alias: string;
    name?: string | null;
};

export type StoreReferenceDto = {
    id: string;
    alias: string;
};
// ...
export type TaxClassDto = {
    id: string;
    alias: string;
    name?: string | null;
    store?: StoreBasicDto | StoreDto | StoreNamedReferenceDto | StoreReferenceDto | null;
    defaultTaxRate?: number | null;
    defaultTaxCode?: string | null;
    countryRegionTaxClasses?: Array<CountryRegionTaxClassDto> | null;
    sortOrder?: number | null;
};

export type TaxClassNamedReferenceDto = {
    id: string;
    alias: string;
    name?: string | null;
};

export type TaxClassReferenceDto = {
    id: string;
    alias: string;
};